Learning Spatio-Temporal Classification and Localization/requirements.txt
absl-py==0.9.0
astor==0.8.1
backcall==0.1.0
cachetools==4.0.0
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
cloudpickle==1.2.2
cycler==0.10.0
Cython==0.29.14
cytoolz==0.10.1
dask==2.9.0
decorator==4.4.1
easydict==1.9
gast==0.2.2
google-auth==1.10.1
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
grpcio==1.26.0
h5py==2.10.0
idna==2.8
imageio==2.6.1
imgaug==0.3.0
ipdb==0.12.3
ipython==7.11.1
ipython-genutils==0.2.0
jedi==0.15.2
joblib==0.14.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.1.1
matplotlib==3.1.1
mkl-fft==1.0.15
mkl-random==1.1.0
mkl-service==2.3.0
networkx==2.4
numpy==1.17.4
oauthlib==3.1.0
olefile==0.46
opt-einsum==3.1.0
pandas==0.25.3
parso==0.5.2
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.2.1
prompt-toolkit==3.0.2
protobuf==3.11.2
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.0
pycparser==2.19
Pygments==2.5.2
pyparsing==2.4.5
python-dateutil==2.8.1
pytz==2019.3
PyWavelets==1.1.1
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
scikit-image==0.15.0
scikit-learn==0.22.1
scikit-video==1.1.11
scipy==1.4.1
Shapely==1.6.4.post2
six==1.13.0
tensorboard==2.1.0
tensorflow==2.1.0
tensorflow-estimator==2.1.0
termcolor==1.1.0
toolz==0.10.0
torch==1.3.1
torchsummary==1.5.1
torchvision==0.4.2
tornado==6.0.3
tqdm==4.40.0
traitlets==4.3.3
unrar==0.4
urllib3==1.25.7
wcwidth==0.1.8
Werkzeug==0.16.0
wrapt==1.11.2
